Refactor large files into smaller modules
